Move manuscript content related components into shared folder.
Plus reduce code redundancy.
