--- src/article/metadata/EntitiesPackage.js
@@ -1,9 +1,0 @@
-import EntityLabelsPackage from './EntityLabelsPackage'
-
-// TODO: remove this
-export default {
-  name: 'entities',
-  configure (config, options = {}) {
-    config.import(EntityLabelsPackage)
-  }
-}
