--- src/article/metadata/index.js
@@ -1,1 +1,0 @@
-export { default as EntitiesPackage } from './EntitiesPackage'
