+++ src/article/shared/AffiliationsListComponent.js
