+++ src/article/shared/DispQuoteComponent.js
