+++ src/article/shared/EditorsListComponent.js
