+++ src/article/shared/EntityLabelsPackage.js
