+++ src/article/shared/FigPreview.js
