+++ src/article/shared/HeadingComponent.js
