+++ src/article/shared/SeparatorComponent.js
