+++ src/article/shared/SigBlockComponent.js
