+++ src/article/shared/TableCellComponent.js
