+++ src/article/shared/TableClipboard.js
