+++ src/article/shared/TableEditing.js
