+++ src/article/shared/TableFigPreview.js
